+++ dev/com.ibm.ws.sipcontainer/src/com/ibm/ws/sip/container/servlets/SipApplicationSessionImpl.java
@@ -1521,7 +1521,7 @@ implements SipApplicationSession {
 	public int extractAppSessionCounter()
 	{
 		if (m_extractedAppSessionSeqCounter < 0) {
-			m_extractedAppSessionSeqCounter = extractAppSessionCounter(getSharedId());
+			m_extractedAppSessionSeqCounter = (int) extractAppSessionCounter(getSharedId());
 		}
 		return m_extractedAppSessionSeqCounter;
 	}
@@ -1537,7 +1537,7 @@ implements SipApplicationSession {
 	 * 6 is the internal Transaction user ID
 	 * @return - the last two digits of the SAS ID.
 	 */
-	public static int extractAppSessionCounter(String AppSessionId)
+	public static long extractAppSessionCounter(String AppSessionId)
 	{
 		StringTokenizer tokenizer = new StringTokenizer(AppSessionId,Replicatable.ID_INTERNAL_SEPERATOR);
 		tokenizer.nextToken(); // skip the serverid
@@ -1545,9 +1545,9 @@ implements SipApplicationSession {
 		if (c_logger.isTraceDebugEnabled()) {
 			c_logger.traceDebug("SipApplicationSessionImpl", "extractAppSessionCounter","found App session counter:"+appCounterAsString);
 		}
-		int result = Integer.parseInt(appCounterAsString);
-		//Moti: fix for defect 487485: apprantly getAppSession().getID().hashcode()
-		// is not univormly distributed (prefix is always the same:logical name).
+		long result =  Long.parseLong(appCounterAsString);
+		//Moti: fix for defect 487485: apparently getAppSession().getID().hashcode()
+		// is not uniformly distributed (prefix is always the same:logical name).
 		// so we will extract the only thing that actually changes: 
 		// the AppSession number. which is the last digit and also
 		// the global app session internal counter
